requirements.txt
Django==4.2.26
Pillow==12.0.0
python-decouple==3.8
psycopg2-binary==2.9.9
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-cors-headers==4.3.1
requests==2.32.5
celery==5.5.3
redis==7.0.1
gunicorn==23.0.0
whitenoise==6.6.0